# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Financial Advisor - Web Search Service",
    "description": "A small FastAPI microservice that provides a unified web search API for the AI Financial Advisor agent. This service should be used for general information lookup that is *not* strictly tied to the Trade Engine's quantitative market data.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/stocks/ranking": {
      "get": {
        "tags": ["stock-ranking"],
        "summary": "Get Stock Ranking",
        "description": "Return stocks ranked by composite score (0-100).",
        "operationId": "get_stock_ranking_api_stocks_ranking_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "description": "Top N stocks to return",
              "title": "Limit"
            },
            "description": "Top N stocks to return"
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100,
              "default": 0.0,
              "description": "Minimum composite score filter",
              "title": "Min Score"
            },
            "description": "Minimum composite score filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RankingResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RankingResponse": {
        "properties": {
          "stocks": {
            "items": { "$ref": "#/components/schemas/StockScore" },
            "type": "array",
            "title": "Stocks"
          },
          "total": { "type": "integer", "title": "Total" },
          "last_ranked_at": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Last Ranked At"
          },
          "data_age_hours": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Data Age Hours"
          }
        },
        "type": "object",
        "required": ["stocks", "total"],
        "title": "RankingResponse"
      },
...
      "StockScore": {
        "properties": {
          "ticker": { "type": "string", "title": "Ticker" },
          "symbol": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Symbol" },
          "name": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Name" },
          "change_percent": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Change Percent" },
          "updated_at": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Updated At" },
          "composite_score": { "type": "number", "title": "Composite Score" },
          "momentum_score": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Momentum Score" },
          "technical_score": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Technical Score" },
          "fundamental_score": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Fundamental Score" },
          "consistency_score": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Consistency Score" },
          "signal_score": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Signal Score" },
          "momentum_1m": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Momentum 1M" },
          "momentum_3m": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Momentum 3M" },
          "momentum_6m": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Momentum 6M" },
          "momentum_12m": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Momentum 12M" },
          "fundamental_trend": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Fundamental Trend" },
          "rank_tier": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Rank Tier" },
          "conviction": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Conviction" },
          "ranked_at": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Ranked At" }
        },
        "type": "object",
        "required": ["ticker", "composite_score"],
        "title": "StockScore"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                { "type": "string" },
                { "type": "integer" }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}